+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 CHANGELOG for 1.x
 ===================
-## v1.0.6
+## v1.1.0 - (2024-03-25)
 ### Added
 - new `ArrayUtils` methods : `checkIssetKeys`, `trimExplode`, `removeEmpty`, `filterByPattern`, `flatToMap`
 - new `MathUtils` methods : `calculateAverage`, `calculateDivision`
-- new `DateUtils` methods : `getFirstDayYearFromDatetime`, `getLastDayMonthFromDatetime`, `getMonthsBetweenDates`, `getDateTimeMonth`, `getDateTimeYear`, `getNbOfWorkingDaysBetweenDatetime`
+- new `DateUtils` methods : `getFirstDayYearFromDateTime`, `getLastDayMonthFromDateTime`, `getMonthsBetweenDates`, `getDateTimeMonth`, `getDateTimeYear`, `getNbOfWorkingDaysBetweenDateTime`, `getDateTimeFromMonthYear`, `getNbDayBetweenDateTimes`, `isNighttime`, `getLastDayMonth`, `getLastDayPreviousMonth`, `getFirstDayNextMonth`, `getFirstDayMonth`, `getNextBirthdayDateTime`, `getFormattedLongMonth`, `getFormattedLongMonthYears`, `getFormattedShortMonthYears`, `addDays`, `subDays`, `addMonths`, `subMonths`, `addYears`, `subYears`
 
 ## v1.0.5 - (2024-03-15)
 ### Added